requirements.txt
opencv-python
numpy
matplotlib
tensorflow-macos
tensorflow-metal
torch
torchvision